backend/Cost_effective_transportation 2/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask